[Quantization] Add metadata porting for nodes added by quantization
Summary:
This diff adds adding metadata to q-dq nodes by inferring the
quatization intent from node annotations. Annotations on the node are
way for user to specify how a node or subgraph is supposed to be
quantized. We continue to use that information to copy metadata on Q/DQ
node from appropriate nodes.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 5f380f81cc4c0eb461fecc75b3ba6ed815a57761
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107107